# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      errors:
        type: object
      message:
        example: error message
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  dto.MaterialAPIResponse:
    properties:
      data:
        $ref: '#/definitions/models.Material'
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.MaterialListAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Material'
        type: array
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.SubMaterialListAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.SubMaterial'
        type: array
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.UpdateStatusRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/models.RecordStatus'
        enum:
        - ACTIVE
        - INACTIVE
        example: ACTIVE
    required:
    - status
    type: object
...
  dto.UpsertMaterialRequest:
    properties:
      description:
        example: Mengenal kendaraan.
        maxLength: 2000
        type: string
      name:
        example: Pengenalan Kendaraan
        maxLength: 200
        minLength: 2
        type: string
      sequence:
        example: 1
        type: integer
    required:
    - name
    - sequence
    type: object
...
info:
  contact: {}
  description: REST API Final Project Bootcamp Intensif Golang
  title: Driving Course Management System API
  version: "1.0"
paths:
...
  /api/v1/admin/materials:
    get:
      description: List curriculum materials. Requires an ACTIVE ADMIN account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MaterialListAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List curriculum materials
      tags:
      - Admin Materials
    post:
      consumes:
      - application/json
      description: Create curriculum material. Requires an ACTIVE ADMIN account.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertMaterialRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MaterialAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create curriculum material
      tags:
      - Admin Materials
  /api/v1/admin/materials/{id}:
    get:
      description: Get curriculum material. Requires an ACTIVE ADMIN account.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MaterialAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get curriculum material
      tags:
      - Admin Materials
    put:
      consumes:
      - application/json
      description: Update curriculum material. Requires an ACTIVE ADMIN account.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertMaterialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MaterialAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update curriculum material
      tags:
      - Admin Materials
  /api/v1/admin/materials/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update material status. Requires an ACTIVE ADMIN account.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MaterialAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update material status
      tags:
      - Admin Materials
  /api/v1/admin/materials/{material_id}/sub-materials:
    get:
      description: List sub-materials. Requires an ACTIVE ADMIN account.
      parameters:
      - description: Resource ID
        in: path
        name: material_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubMaterialListAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List sub-materials
      tags:
      - Admin Sub Materials
    post:
      consumes:
      - application/json
      description: Create sub-material. Requires an ACTIVE ADMIN account.
      parameters:
      - description: Resource ID
        in: path
        name: material_id
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertSubMaterialRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SubMaterialAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create sub-material
      tags:
      - Admin Sub Materials
...
  /api/v1/admin/packages/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update package status. Requires an ACTIVE ADMIN account.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PackageAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update package status
      tags:
      - Admin Packages
...
  /api/v1/admin/students:
    get:
      description: List students. Requires an ACTIVE ADMIN account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StudentListAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List students
      tags:
      - Admin Students
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: 'Enter the token with the Bearer prefix, for example: "Bearer eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"